+++ src/lexer.l
@@ -1,5 +1,5 @@
 %{
-    #include "ast.h"
+    #include "ast_parse.h"
     #include "parser.h"
     #define SAVE_STRING yylval->string = strndup(yytext, yyleng)
     #define SAVE_INTEGER yylval->integer = atoi(yytext)
