+++ tests/performance_tests.c
@@ -4,7 +4,7 @@
 #include <string.h> 
 #include <inttypes.h>
  
-#include "ast.h" 
+#include "ast_parse.h" 
 #include "parser.h" 
 #include "minunit.h"
